requirements.txt
robotframework==7.0.0
robotframework-seleniumlibrary==6.2.0
selenium==4.20.0
numpy ==2.2.6
opencv-python==4.12.0.88
mss ==7.0.1
robotframework-screencaplibrary==1.6.0
robotframework-pabot==5.1.0

